Add files via upload
so this is a test for the logo. we still need  a slogan (catch phrase) of some sort. also we agreed on a blue (i think) but we decided to adde a gradient effect. unfortunately the white in the gradience means it is kinda hard to see. plus me and eloise prefer a orange to yellow gradience, which we will upload in a day or two